Assignment2_raghulch/requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scipy>=1.11
pillow>=10.0
pdf2image>=1.16
reportlab>=4.0
matplotlib>=3.8
tqdm>=4.66
python-docx>=1.1
wikipedia-api>=0.6.0
faker>=24.2.0
PyPDF2>=3.0
